Got a model to train, still having trouble isolating one feature for training
This comment pertains to the code within the "train the model" portion of the code. X is supposed to be the data with one feature dropped or removed from the table.  y is that feature by itself. The model is supposed to predict y based on X. However, I could only get the pipeline to work if I input the entire data set in line 67 i.e.

train_test_split(training_data, y, test_size=0.2) rather than
train_test_split(X, y, test_size=0.2)

Obviously this leads to a perfect model because you can predict y perfectly if you are given y. This is by no means in working order but at least a model is being trained.